+++ src/Routes/StackScreen.js
@@ -26,6 +26,7 @@ import H5PPlayerOffline from '../screens/PlayerScreen/H5PPlayer/H5PPlayerOffline
 import HTMLPlayer from '../screens/PlayerScreen/HTMLPlayer/HTMLPlayer';
 import HTMLPlayerOffline from '../screens/PlayerScreen/HTMLPlayer/HTMLPlayerOffline';
 import YoutubePlayer from '../screens/PlayerScreen/YoutubePlayer/YoutubePlayer';
+import StandAlonePlayer from '../screens/PlayerScreen/StandAlonePlayer/StandAlonePlayer';
 
 const StackScreen = (props) => {
   const Stack = createNativeStackNavigator();
@@ -186,6 +187,13 @@ const StackScreen = (props) => {
       <Stack.Screen
         name="YoutubePlayer"
         component={YoutubePlayer}
+        options={{
+          headerShown: true,
+        }}
+      />
+      <Stack.Screen
+        name="StandAlonePlayer"
+        component={StandAlonePlayer}
         options={{
           headerShown: true,
         }}
