+++ src/screens/PlayerScreen/StandAlonePlayer/data.js
@@ -1,0 +1,198 @@
+export const qumlPlayerConfig = {
+  context: {
+    threshold: 3,
+    mode: 'play',
+    authToken: ' ',
+    sid: '913b3c6c-2874-26dd-ed0c-c23ddc00b718',
+    did: '561c348e631fd225b46a5571cbd42ad1',
+    uid: '',
+    channel: '01268904781886259221',
+    pdata: {
+      id: 'preprod.diksha.portal',
+      ver: '3.3.0',
+      pid: 'sunbird-portal.contentplayer',
+    },
+    contextRollup: {
+      l1: 'string',
+      l2: 'string',
+      l3: 'string',
+      l4: 'string',
+    },
+    tags: [],
+    cdata: [
+      {
+        id: 'c0c9384a82a75f219468d363e1891963',
+        type: 'ContentSession',
+      },
+      {
+        id: 'a12f45a1d7078901adb27b48be4b428d',
+        type: 'PlaySession',
+      },
+    ],
+    timeDiff: 5,
+    objectRollup: {
+      l1: 'string',
+      l2: 'string',
+      l3: 'string',
+      l4: 'string',
+    },
+    host: '',
+    endpoint: '/data/v3/telemetry',
+    userData: {
+      firstName: 'Diptesh',
+      lastName: 'Mukherjee',
+    },
+  },
+  metadata: {},
+  config: {
+    traceId: '1234',
+    sideMenu: {
+      showShare: false,
+      showDownload: false,
+      showExit: false,
+      showPrint: false,
+      showReplay: true,
+    },
+  },
+};
+
+export const questionsData = {
+  questions_data: {},
+};
+
+export const contentPlayerConfig = {
+  context: {},
+  config: { showEndPage: true },
+  metadata: {},
+  data: {},
+};
+
+export const pdfPlayerConfig = {
+  context: {
+    mode: 'play',
+    partner: [],
+    pdata: {
+      id: 'dev.sunbird.portal',
+      ver: '5.2.0',
+      pid: 'sunbird-portal',
+    },
+    contentId: 'do_21374910251798528014586',
+    sid: '6d1898db-d783-4f83-8b92-4a36636e0d2f',
+    uid: 'anonymous',
+    timeDiff: -0.089,
+    channel: '01269878797503692810',
+    tags: ['01269878797503692810'],
+    did: '3ca74a4c5fbce6b7b7f5cd12cebb1682',
+    contextRollup: { l1: '01269878797503692810' },
+    objectRollup: {},
+    userData: { firstName: 'Guest', lastName: '' },
+  },
+  config: {
+    showEndPage: false,
+    endPage: [{ template: 'assessment', contentType: ['SelfAssess'] }],
+    showStartPage: true,
+    host: '',
+    overlay: { showUser: false },
+    splash: {
+      text: '',
+      icon: '',
+      bgImage: 'assets/icons/splacebackground_1.png',
+      webLink: '',
+    },
+    apislug: '/action',
+    repos: ['/sunbird-plugins/renderer'],
+    plugins: [
+      { id: 'org.sunbird.iframeEvent', ver: 1, type: 'plugin' },
+      { id: 'org.sunbird.player.endpage', ver: 1.1, type: 'plugin' },
+    ],
+    sideMenu: {
+      showShare: false,
+      showDownload: false,
+      showExit: false,
+      showPrint: false,
+      showReplay: true,
+    },
+  },
+  metadata: {},
+  data: {},
+};
+
+export const videoPlayerConfig = {
+  context: {
+    mode: 'play',
+    authToken: '',
+    sid: '7283cf2e-d215-9944-b0c5-269489c6fa56',
+    did: '3c0a3724311fe944dec5df559cc4e006',
+    uid: 'anonymous',
+    channel: '505c7c48ac6dc1edc9b08f21db5a571d',
+    pdata: {
+      id: 'prod.diksha.portal',
+      ver: '3.2.12',
+      pid: 'sunbird-portal.contentplayer',
+    },
+    contextRollup: { l1: '505c7c48ac6dc1edc9b08f21db5a571d' },
+    tags: [''],
+    cdata: [],
+    timeDiff: 0,
+    objectRollup: {},
+    host: '',
+    endpoint: '',
+    userData: {
+      firstName: 'Harish Kumar',
+      lastName: 'Gangula',
+    },
+  },
+  config: {
+    traceId: 'afhjgh',
+    sideMenu: {
+      showShare: false,
+      showDownload: false,
+      showExit: false,
+      showPrint: false,
+      showReplay: true,
+    },
+  },
+  metadata: {},
+  data: {},
+};
+
+export const epubPlayerConfig = {
+  context: {
+    mode: 'play',
+    authToken: '',
+    sid: '7283cf2e-d215-9944-b0c5-269489c6fa56',
+    did: '3c0a3724311fe944dec5df559cc4e006',
+    uid: 'anonymous',
+    channel: '505c7c48ac6dc1edc9b08f21db5a571d',
+    pdata: {
+      id: 'prod.diksha.portal',
+      ver: '3.2.12',
+      pid: 'sunbird-portal.contentplayer',
+    },
+    contextRollup: { l1: '505c7c48ac6dc1edc9b08f21db5a571d' },
+    tags: [''],
+    cdata: [],
+    timeDiff: 0,
+    objectRollup: {},
+    host: '',
+    endpoint: '',
+    userData: {
+      firstName: 'Vivek',
+      lastName: 'Kasture',
+    },
+  },
+  config: {
+    traceId: '123456',
+    sideMenu: {
+      showShare: false,
+      showDownload: false,
+      showExit: false,
+      showPrint: false,
+      showReplay: true,
+    },
+  },
+  metadata: {},
+  data: {},
+};
+
+
